created package trackrecycleview and moved utills
