+++ app/src/main/java/com/example/playlistmarker/trackrecyclerview/Track.kt
@@ -1,3 +1,3 @@
-package com.example.playlistmarker
+package com.example.playlistmarker.trackrecyclerview
 
 data class Track (val trackName: String, val artistName: String, val trackTime: String, val artworkUrl100: String)
