+++ app/src/main/java/com/example/playlistmarker/trackrecyclerview/TrackViewHolder.kt
@@ -1,14 +1,14 @@
-package com.example.playlistmarker
+package com.example.playlistmarker.trackrecyclerview
 
 import android.content.Context
 import android.util.TypedValue
-import android.view.RoundedCorner
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.example.playlistmarker.R
 
 class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
